Refactored login details into own module
